### 2022-11-21 #4948 @Kevin-Mattheus-Moerman
@PaulRitsche thanks for this submission. I am the AEiC for this track and also the handling editor for your submission. I have a quick look at your repository and have the following points for you to work on for now:

- [ ] The "user instructions" https://github.com/PaulRitsche/DLTrack/docs/usage produce a 404 error
- [ ] The test instructions link on the readme produces a 404 error: https://github.com/PaulRitsche/DLTrack/tests
- [ ] [The documentation](https://dltrack.readthedocs.io/en/latest/modules.html#documentation) does not render properly for me (Firefox on Ubuntu) see screenshot:
![Screenshot from 2022-11-21 09-24-34](https://user-images.githubusercontent.com/8392709/203014176-1e072d1c-c1ff-4cca-a4b8-01cb16931a68.png)
Given the above perhaps scan through your readme and make sure all links/instructions/documetation/testing are working.

- [ ] I had a quick look at the paper. Since this submission features image processing, I recommend including some images of the software capabilities in the paper. Perhaps you could consider showing some of the ultrasound images, and some overlaid analysis, or graphs too for instance, or perhaps screenshots of the GUI.  I would recommend adding some such images/visualizations to your readme too.
