+++ docs/source/DL_Track.rst
@@ -1,5 +1,5 @@
 DL\_Track package
------------------
+=================
 
 Subpackages
 -----------
