+++ Paper/paper.md
@@ -38,10 +38,15 @@ Ultrasonography can be used to assess muscle architectural parameters during sta
 
 `DL_Track` is a python package to automatically segment and analyse muscle architectural parameters in longitudinal ultrasonography images and videos of human lower limb muscles. The use of ultrasonography to assess muscle morphological and architectural parameters is increasing in different scientific fields [@naruse2022]. This is due to the high portability, cost-effectiveness and patient-friendlyness compared to other imaging modalities such as MRI. Muscle architectural parameters such as muscle thickness, fascicle length and pennation angle are used to assess muscular adaptations to training, disuse and ageing, especially in the lower limbs [@sarto2021]. Albeit many disadvantages such as subjectivity, time expensiveness and required expertise, muscle architectural parameters in ultrasonography images are most often analyzed manually. A potential reason is the lack of versatile, accessible and easy to use analysis tools. Several algorithms to analyze muscle architectural parameters have been developed within the last years [@cronin2011; @rana2009; @marzilger2018; @drazan2019; @farris2016; @seynnes2020]. Nonetheless, most algorithms only partly automate the analysis, introducing some subjectivity in the analysis process. Moreover, many analyse only single architectural parameters and can be exclusively used for image or video analysis. Most existing methods further rely on hardcoded image filtering processes developed on few example images. Given that image parameters do no suit the designed filters of the if the filters are improperly adjusted, these filtering processes likely fail. On top of that, most previously proposed algorithms lack in user friendliness as the provided code and usage documentation is limited.
 
-`DL_Track` incorporates analysis modalities for fully automated analysis of ultrasonography images. `DL_Track` employs convolutional neural networks (CNNs) for the semantic segmentation of muscle fascicle fragments and aponeuroses. The employed CNNs consist of a VGG16 encoder path pre-trained on ImageNet [@simonyan2015; @deng2009] and a U-net decoder [@ronneberger2015]. Based on the semantic segmentation, muscle thickness, fascicle length and pennation angle are calculated as the distance between the detected aponeuroses, the intersection between extrapolated fascicle trajectories and aponeuroses and the intersection angle between extrapolated fascicle trajectories and detected deep aponeuroses, respectively. The workflow of the `DL_Track` analysis algorithm is demonstrated in \autoref{fig:1}.
+`DL_Track` incorporates analysis modalities for fully automated analysis of ultrasonography images. `DL_Track` employs convolutional neural networks (CNNs) for the semantic segmentation of muscle fascicle fragments and aponeuroses. The employed CNNs consist of a VGG16 encoder path pre-trained on ImageNet [@simonyan2015; @deng2009] and a U-net decoder [@ronneberger2015]. Based on the semantic segmentation, muscle thickness, fascicle length and pennation angle are calculated as the distance between the detected aponeuroses, the intersection between extrapolated fascicle trajectories and aponeuroses and the intersection angle between extrapolated fascicle trajectories and detected deep aponeuroses, respectively. An example of the analysis process of two muscles included in our dataset (vastus lateralis and gastrocnemius medialis) is demonstrated in \autoref{fig:1}. 
 
-![Workflow of the DL_Track analysis algorithm.\label{fig:1}](figure1.png)
+![Process from original input image to output result for images of two muscles, gastrocnemius medialis (GM) and vastus lateralis (VL). Subsequent to inputting the original images into the models, predictions are generated by the models for the aponeuroses (apo) and fascicles as displayed in the binary images. Based on the binary image, the output result is calculated by post-processing operations, fascicles and aponeuroses are drawn and the values for fascicle length, pennation angle and muscle thickness are displayed.\label{fig:1}](figure1.png)
 
[…]
