### 2022-11-24 #4948 @PaulRitsche
@Kevin-Mattheus-Moerman thanks for the suggestions.

- We have fixed the links.
- The documentation link for the code documentation now points directly to the docstrings not the overview of packages and modules. However, displaying the links to each module and not directly the docstrings was intentional, as we thought this is easier to navigate.
- We have added a figure containig the image analysis process and a screenshot of the GUI to the paper and the Readme.md.
- We have substituted figure 1 (Workflow) in the paper with the figure of the image analysis process.

Please let us know in case we need to adapt something else or if our changes were unsufficient.
