### 2022-12-06 #4948 @VDB-Bram
Thanks for the invitation to review this code, @Kevin-Mattheus-Moerman. Though, I will not be able to review it, my apologies. Ultra sound is outside my field of expertise.
