### 2022-12-14 #4948 @Kevin-Mattheus-Moerman
@BKillen05 @MaartenAfschrift @mh-skjelvareid @lassoan @thewtex @wspr @antoinevacavant @deepakri201 @luuleitner  would you be interested in reviewing this submission _DL_Track: a python package to analyse muscle ultrasonography images_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

JOSS reviews focus on the evaluation of [the software](https://github.com/PaulRitsche/DLTrack) and the review of a [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04948/joss.04948/10.21105.joss.04948.pdf).

This is a pre-review issue where reviewers are assigned. Once we have sufficient reviewers we will open a review issue, so the review will take place on GitHub. Here is an example of a complete review if you are curious as to what the process is like:
https://github.com/openjournals/joss-reviews/issues/4859
As you can see, a checklist guides each reviewer through the process.

Let me know if you can help or if you have any more questions. Note we can be quite flexible regarding the time needed to complete the review. Thanks!
