### 2023-01-11 #4948 @BKillen05
Hi Kevin, Sorry for the dealy on this:  A postdoc from my lab is willing it review – CC’d and github a/c WannesSwinnen Bryce A Killen, BExSc Hons PhDPostdoctoral Research FellowHuman Movement Biomechanics Research GroupKU Leuven, Leuven, BelgiumTwitter: @bryce_killen From: Kevin Mattheus MoermanSent: dinsdag 10 januari 2023 12:21To: openjournals/joss-reviewsCc: BKillen05; MentionSubject: Re: [openjournals/joss-reviews] [PRE REVIEW]: DL_Track: a python package to analyse muscle ultrasonography images (Issue ***@***.*** @BKillen05 @MaartenAfschrift @lassoan @thewtex @wspr @antoinevacavant @luuleitner @jcfr @waltsims would you be interested in reviewing this submission DL_Track: a python package to analyse muscle ultrasonography images for JOSS (the Journal of Open Source Software)?JOSS reviews focus on the evaluation of the software and the review of a short paper.This is a pre-review issue where reviewers are assigned. Once we have sufficient reviewers we will open a review issue, so the review will take place on GitHub. Here is an example of a complete review if you are curious as to what the process is like:#4859As you can see, a checklist guides each reviewer through the process.Let me know if you can help or if you have any more questions. Note we can be quite flexible regarding the time needed to complete the review. Thanks!—Reply to this email directly, view it on GitHub, or unsubscribe.You are receiving this because you were mentioned.Message ID: ***@***.***>
