### 2023-01-11 #4948 @Kevin-Mattheus-Moerman
Thanks @WannesSwinnen for being able to review for JOSS. Could you please share your full name and affiliation please? Also I would recommend adding this information to your GitHub profile. Thanks.
