### 2023-01-13 #4948 @WannesSwinnen
Hi Kevin,

Sorry for my late reply, but I feel that I had some miscommunication with
Bryce.
I thought he was looking for someone with ultrasound data who was willing
to compare the new algorithm with results obtained from published
algorithms.
Unfortunately, I have little experience with either GitHub nor Python and
hence I'm afraid I won't have very useful contributions to improve the
algorithm/method or review the manuscript.

I'm sorry for that,
Wannes

On Wed, Jan 11, 2023 at 1:11 PM Kevin Mattheus Moerman <
***@***.***> wrote:

> Thanks @WannesSwinnen <https://github.com/WannesSwinnen> for being able
> to review for JOSS. Could you please share your full name and affiliation
> please? Also I would recommend adding this information to your GitHub
> profile. Thanks.
>
