### 2023-01-16 #4948 @modenaxe
> @modenaxe would you reconsider if we say you can pick this up in the new year?

Hi @Kevin-Mattheus-Moerman I am afraid I am not able to do this, unfortunately. Ultrasonography is not exactly my field so I am not able to suggest an alternative reviewer either...sorry about that!
