### 2023-02-13 #4948 @PaulRitsche
> @PaulRitsche you may suggest them and mention their GitHub handle, but please leave out the @ symbol so they are not tagged. Thanks. Apologies for the delay encountered here so far. I will resume my search again in the coming week.

Thanks for the information and don't worry about the delay.
Here are the names and Github handles:
- Philipp Wirth: philippmwirth
- Giacomo Valli: GiacomoValliPhD
- Antonio Morales: ajmoralesa
- Tim van der Zee: timvanderzee
