### 2023-02-18 #4948 @Kevin-Mattheus-Moerman
@GiacomoValliPhD thanks for the comments. I'm unsure about whether this would be a conflict of interest. Are you able to clarify how you know him? Or how that might be (or be perceived as) a conflict of interest? Thanks
