### 2023-02-19 #4948 @GiacomoValliPhD
@Kevin-Mattheus-Moerman I met @PaulRitsche when he came to my university more than a year ago to collaborate with some colleagues that were working on ultrasound imaging for the development of DeepACSA. I was not working with ultrasound so I was not included in the development and publication of DeepACSA. However, we remained in touch to exchange feedback and support on coding.

I have been reviewing the JOSS COIs "_As a reviewer (or editor), COIs are your present or previous association with any authors of a submission: recent (past four years) collaborators in funded research or work that is published; and lifetime for the family members, business partners, and thesis student/advisor or mentor. In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution_" and I don't perceive to fit in any of them.

However, I await your opinion on that.

Best
