+++ DL_Track/gui_helpers/do_calculations_video.py
@@ -2,7 +2,7 @@
 Description
 -----------
 This module contains functions to caculate muscle architectural
-parameters based on binary segmentations by convolutional neural networks (CNNs).
+parameters based on binary segmentations by convolutional neural networks.
 The parameters include muscle thickness, pennation angle and fascicle length.
 First, input images are segmented by the CNNs. Then the predicted aponeuroses
 and fascicle fragments are thresholded and filtered. Fascicle fragments
@@ -92,13 +92,16 @@ def doCalculationsVideo(
         This can be "no_flip" (video is not flipped) or "flipe"
         (video is flipped).
     apo_modelpath : str
-        String variable containing the absolute path to the aponeurosis neural network.
+        String variable containing the absolute path to the aponeurosis
+        neural network.
     fasc_modelpath : str
-        String variable containing the absolute path to the fascicle neural network.
+        String variable containing the absolute path to the fascicle
+        neural network.
     calib_dist : int
         Integer variable containing the distance between the two
         specified point in pixel units. The points must be 10mm
-        apart. Must be non-negative. If "None", the values are outputted in pixel units.
+        apart. Must be non-negative. If "None", the values are outputted in
+        pixel units.
     dic : dict
         Dictionary variable containing analysis parameters.
         These include must include apo_threshold, fasc_threshold,
@@ -147,7 +150,9 @@ def doCalculationsVideo(
                         apo_modelpath="C:/Users/admin/Documents/DL_Track/Models_DL_Track/Final_models/model-VGG16-fasc-BCE-512.h5",
                         fasc_modelpath="C:/Users/admin/Documents/DL_Track/Models_DL_Track/Final_models/model-apo-VGG-BCE-512.h5",
                         calib_dist=98,
-                        dic={'apo_treshold': '0.2', 'fasc_threshold': '0.05', 'fasc_cont_thresh': '40', 'min_width': '60', 'min_pennation': '10', 'max_pennation': '40'},
+                        dic={'apo_treshold': '0.2', 'fasc_threshold': '0.05',
+                        'fasc_cont_thresh': '40', 'min_width': '60',
+                        'min_pennation': '10', 'max_pennation': '40'},
                         gui=<__main__.DLTrack object at 0x000002BFA7528190>)
     [array([60.5451731 , 58.86892027, 64.16011534, 55.46192704, 63.40711356]), ..., array([64.90849385, 60.31621836])]
     [[19.124207107383114, 19.409753216521565, 18.05706763600641, 20.54453899050867, 17.808652286488794], ..., [17.26241882195032, 16.284803480359543]]
@@ -162,7 +167,7 @@ def doCalculationsVideo(
             if float(value) <= 0:
                 tk.messagebox.showerror(
                     "Information",
-                    "Analysis paremters must be non-zero" + " and non-negative",
+                    "Analysis paremters must be non-zero" + " and non-negative"
                 )
                 gui.should_stop = False
                 gui.is_running = False
@@ -206,7 +211,8 @@ def doCalculationsVideo(
             h = img.shape[0]
             w = img.shape[1]
             img = np.reshape(img, [-1, h, w, 3])
-            img = resize(img, (1, 512, 512, 3), mode="constant", preserve_range=True)
+            img = resize(img, (1, 512, 512, 3), mode="constant",
+                         preserve_range=True)
             img = img / 255.0
 
             # Predict aponeurosis
@@ -231,7 +237,7 @@ def doCalculationsVideo(
             pred_fasc_t = resize(pred_fasc_t, (1, h, w, 1))
             pred_fasc_t = np.reshape(pred_fasc_t, (h, w))
 
-            ## Aponuerosis calculation PArt
[…]
