+++ DL_Track/DLTrack_GUI.py
@@ -1,19 +1,23 @@
 """
 Description
 -----------
-This module contains a class with methods to automatically and manually annotate
-longitudinal ultrasonography images and videos. When the class is initiated,
-a graphical user interface is opened. This is the main GUI of the DL_Track package.
+This module contains a class with methods to automatically and manually
+annotate longitudinal ultrasonography images and videos. When the class
+is initiated,a graphical user interface is opened.
+This is the main GUI of the DL_Track package.
 From here, the user is able to navigate all functionalities of the package.
-These extend the methods in this class. The main functionalities of the GUI contained
-in this module are automatic and manual evalution of muscle ultrasonography images.
+These extend the methods in this class. The main functionalities of the GUI
+contained in this module are automatic and manual evalution of muscle
+ultrasonography images.
 Inputted images or videos are analyzed and the parameters muscle fascicle
-length, pennation angle and muscle thickness are returned for each image or video frame.
+length, pennation angle and muscle thickness are returned for each image
+or video frame.
 The parameters are analyzed using convolutional neural networks (U-net, VGG16).
-This module and all submodules contained in the /gui_helpers modality are extensions
-and improvements of the work presented in Cronin et al. (2020). There, the core functionalities
-of this code are already outlined and the comparability of the model segmentations to
-manual analysis (current gold standard) is described. Here, we have improved the code by
+This module and all submodules contained in the /gui_helpers modality
+are extensions and improvements of the work presented in Cronin et al. (2020).
+There, the core functionalities of this code are already outlined and the
+comparability of the model segmentations to manual analysis (current gold
+standard) is described. Here, we have improved the code by
 integrating everything into a graphical user interface.
 
 Functions scope
@@ -27,9 +31,9 @@ tutorials provided for this package.
 
 References
 ----------
-VGG16: Simonyan, Karen, and Andrew Zisserman. “Very deep convolutional networks for large-scale image recognition.” arXiv preprint arXiv:1409.1556 (2014)
-U-net: Ronneberger, O., Fischer, P. and Brox, T. "U-Net: Convolutional Networks for Biomedical Image Segmentation." arXiv preprint arXiv:1505.04597 (2015)
-DL_Track: Cronin, Neil J. and Finni, Taija and Seynnes, Olivier. "Fully automated analysis of muscle architecture from B-mode ultrasound images with deep learning." arXiv preprint arXiv:https://arxiv.org/abs/2009.04790 (2020)
+[1] VGG16: Simonyan, Karen, and Andrew Zisserman. “Very deep convolutional networks for large-scale image recognition.” arXiv preprint arXiv:1409.1556 (2014)
+[2] U-net: Ronneberger, O., Fischer, P. and Brox, T. "U-Net: Convolutional Networks for Biomedical Image Segmentation." arXiv preprint arXiv:1505.04597 (2015)
+[3] DL_Track: Cronin, Neil J. and Finni, Taija and Seynnes, Olivier. "Fully automated analysis of muscle architecture from B-mode ultrasound images with deep learning." arXiv preprint arXiv:https://arxiv.org/abs/2009.04790 (2020)
 """
 
 import tkinter as tk
@@ -45,8 +49,8 @@ class DLTrack:
     ultrasonography images/videos of human lower limb muscles.
     An analysis tkinter GUI is opened upon initialization of the class.
     By clicking the buttons, the user can switch between different
-    analysis modes for image/video analysis and model training. The GUI consists
-    of the following elements.
+    analysis modes for image/video analysis and model training.
+    The GUI consists of the following elements.
     - Input Directory:
     By pressing the "Input" button, the user is asked to select
[…]
