+++ Paper/paper.bib
@@ -223,3 +223,4 @@ Conclusion  The proposed algorithm provides high measurement reliability. Howe
 	pages = {291--301},
 	file = {Marzilger et al. - 2018 - Reliability of a semi-automated algorithm for the .pdf:C\:\\Users\\ritpau00\\Zotero\\storage\\N6V6GE2L\\Marzilger et al. - 2018 - Reliability of a semi-automated algorithm for the .pdf:application/pdf},
 }
+
