+++ Paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'DL_Track: a python package to analyse muscle ultrasonography images'
+title: 'DL_Track_US: a python package to analyse muscle ultrasonography images'
 tags:
   - Python
   - muscle
@@ -32,20 +32,21 @@ bibliography: paper.bib
 
 # Summary
 
-Ultrasonography can be used to assess muscle architectural parameters during static and dynamic conditions. Nevertheless, the analysis of the acquired ultrasonography images presents a major difficulty. Muscle architectural parameters such as muscle thickness, fascicle length and pennation angle are mainly segmented manually. Manual analysis is time expensive, subjective and requires thorough expertise. Within recent years, several algorithms were developed to solve these issues. Yet, these are only partly automated, are not openly available, or lack in user friendliness. The DL_Track python package is designed to allow fully automated and rapid analyis of muscle architectural parameters in lower limb ultrasonography images.
+Ultrasonography can be used to assess muscle architectural parameters during static and dynamic conditions. Nevertheless, the analysis of the acquired ultrasonography images presents a major difficulty. Muscle architectural parameters such as muscle thickness, fascicle length and pennation angle are mainly segmented manually. Manual analysis is time expensive, subjective and requires thorough expertise. Within recent years, several algorithms were developed to solve these issues. Yet, these are only partly automated, are not openly available, or lack in user friendliness. The DL_Track_US python package is designed to allow fully automated and rapid analyis of muscle architectural parameters in lower limb ultrasonography images.
 
 # Statement of need
 
-`DL_Track` is a python package to automatically segment and analyse muscle architectural parameters in longitudinal ultrasonography images and videos of human lower limb muscles. The use of ultrasonography to assess muscle morphological and architectural parameters is increasing in different scientific fields [@naruse2022]. This is due to the high portability, cost-effectiveness and patient-friendlyness compared to other imaging modalities such as MRI. Muscle architectural parameters such as muscle thickness, fascicle length and pennation angle are used to assess muscular adaptations to training, disuse and ageing, especially in the lower limbs [@sarto2021]. Albeit many disadvantages such as subjectivity, time expensiveness and required expertise, muscle architectural parameters in ultrasonography images are most often analyzed manually. A potential reason is the lack of versatile, accessible and easy to use analysis tools. Several algorithms to analyze muscle architectural parameters have been developed within the last years [@cronin2011; @rana2009; @marzilger2018; @drazan2019; @farris2016; @seynnes2020]. Nonetheless, most algorithms only partly automate the analysis, introducing some subjectivity in the analysis process. Moreover, many analyse only single architectural parameters and can be exclusively used for image or video analysis. Most existing methods further rely on hardcoded image filtering processes developed on few example images. Given that image parameters do no suit the designed filters of the if the filters are improperly adjusted, these filtering processes likely fail. On top of that, most previously proposed algorithms lack in user friendliness as the provided code and usage documentation is limited.
[…]
