+++ DL_Track_US/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["__version__", "DL_Track_US_GUI"]
+
+__version__ = "0.1.1"
+
+__author__ = ["Paul Ritsche, Olivier Seynnes, Neil Cronin"]
+
+from DL_Track_US.DL_Track_US_GUI import *
