+++ Paper/paper.bib
@@ -209,10 +209,10 @@ Conclusions: We have developed a novel automatic fascicle tracking algorithm tha
 	issn = {1439-6319, 1439-6327},
 	url = {http://link.springer.com/10.1007/s00421-017-3769-8},
 	doi = {10.1007/s00421-017-3769-8},
-	abstract = {Purpose  The assessment of muscle architecture with B-mode ultrasound is an established method in muscle physiology and mechanics. There are several manual, semi-automated and automated approaches available for muscle architecture analysis from ultrasound images or videos. However, most approaches have limitations such as workload, subjectivity, drift or they are applicable to short muscle fascicles only. Addressing these issues, an algorithm was developed to analyse architectural parameters of the vastus lateralis muscle (VL).
-Methods  In 17 healthy young men and women, ultrasound images were taken five times on two different days during passive knee joint flexion. From the images, fascicle length (FL), pennation angle (PAN) and muscle thickness (MTH) were calculated for both test days using the algorithm. Interday differences were determined using a two-way ANOVA. Interday and intraday reliability were assessed using intraclass correlation coefficients (ICC) and root mean square (RMS) differences.
-Results  FL, MTH and PAN did not differ between day one and two. The within day ICCs were above 0.94 for all tested parameters. The average interday ICCs were 0.86 for the FL, 0.96 for MTH and 0.60 for PAN. The average RMS differences between both days were 5.0\%, 8.5\% and 12.0\% for MTH, FL and PAN, respectively.
-Conclusion  The proposed algorithm provides high measurement reliability. However, the interday reliability might be influenced by small differences in probe position between days.},
+	abs:tract = {Purpose: The assessment of muscle architecture with B-mode ultrasound is an established method in muscle physiology and mechanics. There are several manual, semi-automated and automated approaches available for muscle architecture analysis from ultrasound images or videos. However, most approaches have limitations such as workload, subjectivity, drift or they are applicable to short muscle fascicles only. Addressing these issues, an algorithm was developed to analyse architectural parameters of the vastus lateralis muscle (VL).
+Methods In 17 healthy young men and women, ultrasound images were taken five times on two different days during passive knee joint flexion. From the images, fascicle length (FL), pennation angle (PAN) and muscle thickness (MTH) were calculated for both test days using the algorithm. Interday differences were determined using a two-way ANOVA. Interday and intraday reliability were assessed using intraclass correlation coefficients (ICC) and root mean square (RMS) differences.
+Results: FL, MTH and PAN did not differ between day one and two. The within day ICCs were above 0.94 for all tested parameters. The average interday ICCs were 0.86 for the FL, 0.96 for MTH and 0.60 for PAN. The average RMS differences between both days were 5.0\%, 8.5\% and 12.0\% for MTH, FL and PAN, respectively.
+Conclusion: The proposed algorithm provides high measurement reliability. However, the interday reliability might be influenced by small differences in probe position between days.},
 	language = {en},
 	number = {2},
 	urldate = {2022-11-08},
@@ -224,3 +224,17 @@ Conclusion  The proposed algorithm provides high measurement reliability. Howe
 	file = {Marzilger et al. - 2018 - Reliability of a semi-automated algorithm for the .pdf:C\:\\Users\\ritpau00\\Zotero\\storage\\N6V6GE2L\\Marzilger et al. - 2018 - Reliability of a semi-automated algorithm for the .pdf:application/pdf},
 }
 
+@article{ritsche2022,
[…]
