+++ DL_Track_US/gui_helpers/image_quality.py
@@ -1,713 +1,0 @@
-"""
-Description
------------
-This module contains a class to augment image brightness, contrast and noise.
-Moreover, using the class, the quality of images can be evaluated. The quality
-score is calculated based on mean pixel intensity, standard deviation or pixel
-intensity, noise estimation, blur estimation and signal-to-noise ratio.
-These parameters were selected because we believe that they are most
-relevant for the evaluation of ultrasound images.
-The module was specifically designed to be executed from a GUI, but each
-function can be used separately.
-When used from the GUI, only the image evaluation funtions are used. A .xlsx
-file containing the parameters will be saved to a specified directory. This
-is the directory where all images to be evaluated should be.
-
-Functions scope
----------------
-For scope of the functions see class documentation.
-
-Notes
------
-Additional information and usage examples can be found at the respective
-functions docstrings.
-"""
-
-import glob
-import os
-import time
-import cv2
-
-import numpy as np
-import pandas as pd
-from scipy import stats
-from sewar.full_ref import ssim
-from skimage import restoration, util
-
-
-class ImageQuali:
-    """
-    Python class to adapt and assess the quality of images.
-    The functionalities offered by this class can adapt the
-    brightness, contrast and noise of images. Image quality can
-    also be assessed. All instance methods can be used
-    separately but the main instance is 'adaptImages' where images can be
-    augmented and their quality assessed.
-
-    Attributes
-    ----------
-    self.root : str
-        Path to root directory containing the images to be augmented/
-        analyzed.
-    self.out : str
-        Path to output directory where augmented images are saved.
-        Note that the quality assessment .xlsx file is always saved
-        to the root path.
-    self.mlocs : list
-        List variable to store coordinates of mouse clicks. This is
-        only used when images are cropped.
-
-    Methods
-    -------
-    __init__
-        Instance method to initialize class
-    mclick
-         Instance method to detect mouse click coorindates in image.
-    getBoundingBox
-        Instance method to get the user selected bounding box and
-        crop the image
-    saveResults
-        Instance method to save image quality assessment results to
-        excel sheet
-    getSignalToNoise
-        Instance method to calculate the signel to noise ratio of a given image.
-    calculateNoise
-        Instance method to assess image quality/noise in several measures
-    makeImageContrast
-        Instance method to change image contrast/brightness
-    makeImageNoise
-        Instance method to add speckle noise to image
-    makeImageBlur
-        Instance method to blur the input image
-    adaptImages
-        Instance method to apply quality changes and assess qualiy
-
-    Notes
-    -----
-    This class solely consists of instance methods.
-    Each method can be used on its own with the respective input parameters.
-    For more information on the instance methods check the respective
-    docstrings.
-
-    Some of the instance methods are not used when the class is initialized.
-    These include the following ones: mclick, getBoundingBox, makeImageContrast,
-    makeImageBlur, makeImageNoise.
-    Moreover, the self.out instance attribute is initialized to be the same
-    directory as the self.input attribute.
-    """
-
-    def __init__(self, rootpath: str, outpath: str):
-        """
-        Instance method to initialize the ImageQuali class.
-
-        Parameters
-        ----------
-        rootpath : str
-            Path to root directory containing images to be
-            augmented/analyzed.
-        outpath : str
-            Path to output directory where images/results
-            are saved.
-        mlocs : list
-            List variable to store detected mouse click
-            coordinates for bounding box selection.
-        """
-        self.root = rootpath
-        self.out = outpath
-        self.mlocs = []
-
[…]
