+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = DL_Track_US
-version = 0.1.1
+version = 0.1.2
 
 [options]
 packages = DL_Track_US
