### 2023-04-06 #5206 @PaulRitsche
@Kevin-Mattheus-Moerman @GiacomoValliPhD
We have fixed the issues @GiacomoValliPhD pointed out in the review. Thanks for the helpful suggestions. The code, paper, documentation and package are updated, as you can see in the repo and on pypi.
Attached you can find a point to point response on the issues raised by @GiacomoValliPhD. In case of any uncertainty or unsatisfying solution, please let us know.
[Point_to_point_Answers_PR_v1.docx](https://github.com/openjournals/joss-reviews/files/11174480/Point_to_point_Answers_PR_v1.docx)
