### 2023-04-15 #5206 @GiacomoValliPhD
@PaulRitsche I am satisfied with the revision, I believe that some changes will be really useful to the final user. I only have 2 minor observations:

1 - Aggressive is misspelled (agressive) in the README file
2 - In the manuscript some formatting went wrong? For the name "DL_Track_US". And something went wrong with the citation? Please see the figure.

![image](https://user-images.githubusercontent.com/81100252/232195631-8061735f-f294-4da8-8207-b4e58c6507b6.png)
