### 2023-04-21 #5206 @PaulRitsche
@Kevin-Mattheus-Moerman @GiacomoValliPhD I have updated the repository.

1. I corrected the spelling in the Readme.md
2. I adapted the formatting in the paper.md
3. I fixed the citation issue in the paper.md.

Thanks for pointing out these issues, I hope all is now resolved to your satisfaction.
