+++ Paper/paper.md
@@ -47,6 +47,6 @@ All `DL_Track_US` functionalities are embedded in a graphical user interface (GU
 ![Start window of the DL_Track_US graphical user interface. Here, the input and model directories can be specified, the analysis type (image, video, image manual, video manual) selected and the analysis parameters specified. Moreover, users can choose to train their own models based on their own input images using the "Train Model" button to open a seperate window.\label{fig:2}](figure2.png)
 
 Although we included images of the human gastrocnemius medialis, tibialis anterior, soleus and vastus lateralis from four different devices, it is likely the provided pre-trained CNNs fail when images from different muscles or devices are analyzed. However, an option to train CNNs bases on user data is also included in the GUI. Users are thereby enabled to train own CNNs based on own image or video data. The training ultrasonography image data, the pre-trained CNNs, example usage files as well as usage and testing instructions are provided.
-`DL_Track_US` was developed to automatically segment muscle architectural parameters in longitudinal ultrasonography images or videos. Nevertheless, ultrasonography can be used to assess muscle anatomical cross-sectional area (ACSA) as well. We recently published DeepACSA [@ritsche], the counterpiece of DL_Track_US. DeepACSA was developed for automatic segmentation of muscle ACSA in transversal, panoramic ultrasonography images of the human rectus femoris, vastus lateralis and gastrocnemius muscles. 
+`DL_Track_US` was developed to automatically segment muscle architectural parameters in longitudinal ultrasonography images or videos. Nevertheless, ultrasonography can be used to assess muscle anatomical cross-sectional area (ACSA) as well. We recently published DeepACSA [@ritsche2022], the counterpiece of `DL_Track_US`. DeepACSA was developed for automatic segmentation of muscle ACSA in transversal, panoramic ultrasonography images of the human rectus femoris, vastus lateralis and gastrocnemius muscles. 
 
 # References
