### 2023-04-22 f6c8230
Fixed: Coloring & Filetypes
