### 2023-04-26 #5206 @PaulRitsche
@Kevin-Mattheus-Moerman @deepakri201
We have fixed the issues @deepakri201  pointed out in the review. Thanks for the helpful suggestions. The code, zenodo-data and package are updated, as you can see in the repo and on pypi.
Attached you can find a point to point response on the issues raised by @deepakri201 . In case of any uncertainty or unsatisfying solution, please let us know.
[Point_to_point_Answers_PR_v2.docx](https://github.com/openjournals/joss-reviews/files/11332514/Point_to_point_Answers_PR_v2.docx)
