### 2023-04-28 #5206 @Kevin-Mattheus-Moerman
@PaulRitsche given the positive reviews I am delighted to say we can proceed towards acceptance in JOSS. There are several steps now needed before we can proceed. :point_up:. Please can you work on your items. I would prefer if you use ZENODO for the archiving. Let me know when you are done. In the mean time I'll also review your paper now and might come back with some minor issues.
