### 2023-04-28 #5206 @Kevin-Mattheus-Moerman
@PaulRitsche please work on the below as well for your paper:
- [ ] Do not use country acronyms in your affiliations, so please spell out `United Kingdom` for instance.
- [ ] Fix the spelling of `analyis`
- [ ] Check if it would be better to use `software` in place of `softwares`, I am unsure if the pluralization like that is correct.
- [ ] Fix spelling of `seperate`
- [ ] Check if `counterpiece` should be written `counter piece`
