+++ Paper/paper.md
@@ -24,7 +24,7 @@ affiliations:
    index: 2
  - name: University of Jyvaskyla, Faculty of Sport and Health Sciences, Jyvaskyla, Finland
    index: 3
- - name: School of Sport & Exercise, University of Gloucestershire, UK
+ - name: School of Sport & Exercise, University of Gloucestershire, United Kingdom
    index: 4
 date: 08 November 2022
 bibliography: paper.bib
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-Ultrasonography can be used to assess muscle architectural parameters during static and dynamic conditions. Nevertheless, the analysis of the acquired ultrasonography images presents a major difficulty. Muscle architectural parameters such as muscle thickness, fascicle length and pennation angle are mainly segmented manually. Manual analysis is time expensive, subjective and requires thorough expertise. Within recent years, several algorithms were developed to solve these issues. Yet, these are only partly automated, are not openly available, or lack in user friendliness. The DL_Track_US python package is designed to allow fully automated and rapid analyis of muscle architectural parameters in lower limb ultrasonography images.
+Ultrasonography can be used to assess muscle architectural parameters during static and dynamic conditions. Nevertheless, the analysis of the acquired ultrasonography images presents a major difficulty. Muscle architectural parameters such as muscle thickness, fascicle length and pennation angle are mainly segmented manually. Manual analysis is time expensive, subjective and requires thorough expertise. Within recent years, several algorithms were developed to solve these issues. Yet, these are only partly automated, are not openly available, or lack in user friendliness. The `DL_Track_US` python package is designed to allow fully automated and rapid analysis of muscle architectural parameters in lower limb ultrasonography images.
 
 # Statement of need
 
@@ -42,11 +42,11 @@ Ultrasonography can be used to assess muscle architectural parameters during sta
 
 ![Process from original input image to output result for images of two muscles, gastrocnemius medialis (GM) and vastus lateralis (VL). Subsequent to inputting the original images into the models, predictions are generated by the models for the aponeuroses (apo) and fascicles as displayed in the binary images. Based on the binary image, the output result is calculated by post-processing operations, fascicles and aponeuroses are drawn and the values for fascicle length, pennation angle and muscle thickness are displayed.\label{fig:1}](figure1.png)
 
-All `DL_Track_US` functionalities are embedded in a graphical user interface (GUI) to allow easy and intuitive usage. Apart from automated analysis, a manual analysis option is included in the GUI as well. This is provided in case the pre-trained CNNs perform badly on input images or videos and subsequently to avoid switching between softwares. The start window of the GUI is presented in \autoref{fig:2} and more information on usage can be found in our repository.
+All `DL_Track_US` functionalities are embedded in a graphical user interface (GUI) to allow easy and intuitive usage. Apart from automated analysis, a manual analysis option is included in the GUI as well. This is provided in case the pre-trained CNNs perform badly on input images or videos and subsequently to avoid switching between software. The start window of the GUI is presented in \autoref{fig:2} and more information on usage can be found in our repository.
 
-![Start window of the DL_Track_US graphical user interface. Here, the input and model directories can be specified, the analysis type (image, video, image manual, video manual) selected and the analysis parameters specified. Moreover, users can choose to train their own models based on their own input images using the "Train Model" button to open a seperate window.\label{fig:2}](figure2.png)
[…]
