### 2023-05-02 #5206 @PaulRitsche
> @PaulRitsche please work on the below as well for your paper:
>
> * [ ]  Do not use country acronyms in your affiliations, so please spell out `United Kingdom` for instance.
> * [ ]  Fix the spelling of `analyis`
> * [ ]  Check if it would be better to use `software` in place of `softwares`, I am unsure if the pluralization like that is correct.
> * [ ]  Fix spelling of `seperate`
> * [ ]  Check if `counterpiece` should be written `counter piece`

@Kevin-Mattheus-Moerman  I just fixed these raised issues and updated the paper.

1. We fixed the acronyms and spelled out the affiliations
2. We fixed the spelling of analyis to "analysis".
3. Software is the correct term here, so we corrected that.
4. We fixed the spelling of seperate to "separate".
5. Couterpiece was replaced by the correct "counterpart".
