### 2023-05-02 #5206 @PaulRitsche
Thanks @Kevin-Mattheus-Moerman @deepakri201 @GiacomoValliPhD for the helpful suggestions in improving the package. It was a great experience for us publishing in JOSS!
